+++ hikka/version.py
@@ -1,2 +1,2 @@
 """Represents current userbot version"""
-__version__ = (1, 2, 0)
+__version__ = (1, 2, 1)
