+++ termux_requirements.txt
@@ -9,6 +9,5 @@ requests==2.27.1
 aiogram==2.19
 grapheme==0.6.0
 heroku3==5.1.4
-psycopg2-binary==2.9.3
 
 # Python 3.8+
